Moved files to a better structure
