+++ .sequelizerc
@@ -1,0 +1,6 @@
+const path = require('path');
+
+module.exports = {
+  'models-path': path.resolve('database', 'models'),
+  'migrations-path': path.resolve('database', 'migrations')
+};
