+++ assets/images/nether_wart.png
