+++ commands/config.js
@@ -1,6 +1,6 @@
 const { MessageEmbed, MessageActionRow } = require('discord.js');
-const { DataHandler } = require('../database.js');
-const { Data } = require('../data.js')
+const { DataHandler } = require('../classes/database.js');
+const { Data } = require('../classes/data.js')
 
 module.exports = {
 	name: 'config',
